Using JSON API, UI autorefresh and DB update
